backend/requirements.txt
fastapi
langchain
langchain-community
langchain-core
langchain-google-genai
uvicorn
bs4
requests
python-dotenv
langchain-faiss-cpu
